refactor: remove web3, remove core/main separation
remove web3, the lib just provides the getProvider func which can use with web3 or any libary that has support. With web3 removed and transports being added soon, there is no longer a need have a main/core seperation. Bundle size is no longer a concern and optionality is no longer a concern as transport will better support those options.
